feat: enable usage of navigationFilter and preFilterFetch (#54)
* feat(compatibility): add feature for navigation filter and ensure compatibility to new component versions
* fix: make env variables accessible in Nuxt module (#53)

BREAKING CHANGE:
The original fsxa-api class was removed and the new ones FSXAProxyApi and FSXARemoteApi are used. Their have a slightly different, but better, method signatures. For more information, please read the migration guide in the CHANGELOG of the FSXA-API.

Environment variables FSXA_HOST and FSXA_PORT has be configured in your production environments. Make sure to set the variables in your PWA deployments. In testing scenarios, localhost:3000 will be used.